cmTarget: Move sanity checks and computed property access to callers
The GetProperty method is now just accessing contained data, meaning it
can be implemented in cmState.

Remove the cmMakefile context from the signature as a result and remove
the overload with the same signature.

As a side-effect, the CMP0026 warning now gives a backtrace to the
target when issued from a generator expression.